packages structure reorganization
