--- src/main/java/gui/swingui/ByTopicModel.java
@@ -1,34 +1,0 @@
-package gui.swingui;
-
-import javax.swing.*;
-import javax.swing.event.ListDataListener;
-
-/**
- * Created by vaba1010 on 09.01.2017.
- */
-public class ByTopicModel implements ComboBoxModel {
-
-    public void setSelectedItem(Object anItem) {
-
-    }
-
-    public Object getSelectedItem() {
-        return null;
-    }
-
-    public int getSize() {
-        return 0;
-    }
-
-    public Object getElementAt(int index) {
-        return null;
-    }
-
-    public void addListDataListener(ListDataListener l) {
-
-    }
-
-    public void removeListDataListener(ListDataListener l) {
-
-    }
-}
