--- src/main/java/gui/swingui/StatisticsWindow.java
@@ -1,9 +1,0 @@
-package gui.swingui;
-
-import javax.swing.*;
-
-/**
- * Created by PrifizGamer on 15.01.2017.
- */
-public class StatisticsWindow extends JFrame {
-}
