+++ src/main/java/gui/swingui/record/EditRecordWindow.java
@@ -1,11 +1,12 @@
-package gui.swingui;
+package gui.swingui.record;
 
-import controller.AddCommand;
 import controller.Command;
+import controller.EditCommand;
 import datamodel.EmptyTheme;
-import datamodel.Language;
+import datamodel.Record;
 import datamodel.Theme;
 import datamodel.Word;
+import gui.swingui.MainWindow;
 import org.apache.commons.io.FileUtils;
 
 import javax.swing.*;
@@ -14,17 +15,52 @@ import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.io.File;
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.prefs.Preferences;
 
 /**
- * Created by PrifizGamer on 09.01.2017.
+ * Created by vaba1010 on 11.01.2017.
  */
-public class AddRecordWindow extends RecordWindow {
+public class EditRecordWindow extends RecordWindow {
 
+    private Record recordToEdit;
+
+    public EditRecordWindow(MainWindow main, Record record) throws HeadlessException {
+        super(main);
+        this.recordToEdit = record;
+        initUmlauts();
+        initOperations();
+        initLayout();
+        this.pack();
+        this.setVisible(true);
+    }
+
+    @Override
+    protected List<Word> initWords() {
+        return recordToEdit.getWords();
+    }
+
+    @Override
+    protected String getWindowTitle() {
+        return "Dictionary - Edit Record";
+    }
+
+    @Override
+    protected void setSelectedTopic() {
+        if(!recordToEdit.getWords().isEmpty()) {
+            existingTopicsCombo.setSelectedItem(recordToEdit.getWords().get(0).getTheme().getName());
+        }
+    }
+
+    @Override
     protected void initPictureChooser() {
-        final String LAST_USED_FOLDER = ".";
+        pictureLabel.setIcon(
+                new ImageIcon(
+                        new ImageIcon(
+                                new File("pictures", recordToEdit.getPictureName()).getAbsolutePath())
+                                .getImage().getScaledInstance(150, 150, Image.SCALE_SMOOTH)));
+        copiedPictureFile = new File(recordToEdit.getPictureName());
+        final String LAST_USED_FOLDER = new File("pictures", recordToEdit.getPictureName()).getAbsolutePath();
         choosePicture.addActionListener(new ActionListener() {
             @Override
             public void actionPerformed(ActionEvent e) {
@@ -32,6 +68,8 @@ public class AddRecordWindow extends RecordWindow {
                 JFileChooser pictureChooser = new JFileChooser(prefs.get(LAST_USED_FOLDER,
                         new File(".").getAbsolutePath()));
                 pictureChooser.setMultiSelectionEnabled(false);
+
+
                 int returnVal = pictureChooser.showOpenDialog(null);
                 if (returnVal == JFileChooser.APPROVE_OPTION) {
                     pictureFile = pictureChooser.getSelectedFile();
@@ -52,8 +90,8 @@ public class AddRecordWindow extends RecordWindow {
         });
     }
 
-    protected void initSaveOperation(final java.util.List<Word> words) {
-
+    protected void initSaveOperation(final List<Word> words) {
+        // FIXME
         saveButton.addActionListener(new ActionListener() {
             public void actionPerformed(ActionEvent e) {
 
@@ -67,49 +105,19 @@ public class AddRecordWindow extends RecordWindow {
                         word.setTheme(new Theme(existingTopicsCombo.getSelectedItem().toString(), "empty description"));
                     }
                 }
-                if(copiedPictureFile == null || !copiedPictureFile.exists()) {
-                    JOptionPane.showMessageDialog(null, "Please, select the picture");
-                } else {
-                    Command addCommand = new AddCommand(words, copiedPictureFile.getName());
+
+                    Command editCommand = new EditCommand(recordToEdit, words, copiedPictureFile.getName());
                     datamodel.Dictionary dictionary = mainWindow.getDictionary();
                     try {
-                        addCommand.execute(dictionary);
+                        editCommand.execute(dictionary);
                     } catch (IOException ex) {
                         JOptionPane.showMessageDialog(null, ex.getMessage());
                     }
                     mainWindow.setDictionary(dictionary);
                     mainWindow.refresh();
                     dispose();
-                }
+
             }
         });
-    }
-
-    public AddRecordWindow(MainWindow main) throws HeadlessException {
-        super(main);
-        initUmlauts();
-        initOperations();
-        initLayout();
-        this.pack();
-        this.setVisible(true);
-    }
-
-    @Override
-    protected void setSelectedTopic() {
-
-    }
-
-    @Override
-    protected List<Word> initWords() {
-        final java.util.List<Word> words = new ArrayList<Word>(3);
-        words.add(new Word("", Language.ENGLISH, new EmptyTheme()));
-        words.add(new Word("", Language.GERMAN, new EmptyTheme()));
-        words.add(new Word("", Language.RUSSIAN, new EmptyTheme()));
-        return words;
-    }
-
-    @Override
-    protected String getWindowTitle() {
-        return "Dictionary - Add New Record";
     }
 }
